Move parts of 'literal.ts' and 'path.ts' imported in 'typesystem.ts' into reflection dir
