+++ src/reflection/generate.ts
@@ -137,8 +137,8 @@ export * as syntax from "./syntax/syntax";`
     let contents = fs.readFileSync(filePath, "utf8");
     // rewrite scoped import paths
     contents = contents.replace(
-      /from "(..\/)?reflection"/g,
-      `from "edgedb/src/reflection"`
+      /from "(..\/)?reflection(.*)"/g,
+      `from "edgedb/src/reflection$2"`
     );
     contents = contents.replace(/from "@generated\//g, `from "../`);
     const outputDir = path.join(to, "syntax");
