+++ src/syntax/params.ts
@@ -6,7 +6,7 @@ import {
   Cardinality,
   ScalarType,
   ArrayType,
-} from "reflection";
+} from "../reflection";
 import {$expressionify} from "./path";
 
 export type $expr_OptionalParam<Type extends ParamType = ParamType> = {
