+++ src/syntax/update.ts
@@ -18,9 +18,8 @@ import {
   NonArrayMaterialType,
   SomeObjectType,
   BaseType,
-} from "reflection";
+} from "../reflection";
 import _std from "@generated/modules/std";
-import {Movie} from "@generated/modules/default";
 import {scalarAssignableBy} from "@generated/castMaps";
 
 /////////////////
